Reorganise mapping code
Group row and channel mapper together
