+++ africanus/averaging/new_averager.py
@@ -9,8 +9,7 @@ from collections import namedtuple
 from numba import types
 import numpy as np
 
-from africanus.averaging.row_mapping import row_mapper
-from africanus.averaging.channel_mapping import channel_mapper
+from africanus.averaging.new_averager_mapping import row_mapper, channel_mapper
 from africanus.util.numba import is_numba_type_none, generated_jit, njit
