+++ africanus/averaging/tests/test_mapping.py
@@ -9,8 +9,7 @@ from numpy.testing import assert_array_equal
 import pytest
 
 from africanus.averaging.support import unique_time, unique_baselines
-from africanus.averaging.row_mapping import row_mapper
-from africanus.averaging.channel_mapping import channel_mapper
+from africanus.averaging.new_averager_mapping import row_mapper, channel_mapper
 
 
 @pytest.fixture
